Refactor folder structure
Loosely followed this recommendation from this Medium article (https://medium.com/habilelabs/best-folder-structure-for-react-native-project-a46405bdba7).
