+++ frontend/src/components/atoms/log_detail/LogDetailBottomSheet.tsx
