+++ frontend/src/components/molecules/logs_list/LogsList.tsx
