+++ frontend/src/screens/LogsView.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import LogsViewHome from '../log_view_home/LogsViewHome';
+import LogsViewHome from '../components/molecules/log_view_home/LogsViewHome';
 
 export type LogsViewProps = {
   data: { title: string; content: string; categories: string[] }[];
