+++ frontend/src/screens/TabView.tsx
@@ -3,7 +3,7 @@ import { SafeAreaView } from 'react-native-safe-area-context';
 import { Text } from 'react-native';
 import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
-import LogsView from '../logs_view/LogsView';
+import LogsView from './LogsView';
 
 const TabView = () => {
